TASK: Move methods `set`, `get`, `getByTag`, `has`, and `remove` from `FrontendInterface` into new `FlowCacheFrontendInterface`
The frontend interface now only describes method needed by the cache manager and the methods used for interaction with the cache
are moved to the new `FlowCacheFrontendInterface`. This change will allow to implement frontends that provide a differrent api like PSR6, PSR16
directly.

About breakiness:
- Since code always use one of the concrete frontend implementations no changed are needed
- Cache Backends still get the same Frontend set and as long as the backends only use the getIdentifier
  method which is common no changes are needed.
